appython/requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
werkzeug==2.0.1